+++ lib/services/local/services_test.go
@@ -60,18 +60,21 @@ func setupServicesContext(ctx context.Context, t *testing.T) *servicesContext {
 	eventsService := NewEventsService(tt.bk)
 	presenceService := NewPresenceService(tt.bk)
 
+	caService := NewCAService(tt.bk)
+
 	tt.suite = &suite.ServicesTestSuite{
-		CAS:           NewCAService(tt.bk),
-		PresenceS:     presenceService,
-		ProvisioningS: NewProvisioningService(tt.bk),
-		WebS:          NewTestIdentityService(tt.bk),
-		Access:        NewAccessService(tt.bk),
-		EventsS:       eventsService,
-		ChangesC:      make(chan interface{}),
-		ConfigS:       configService,
-		LocalConfigS:  configService,
-		RestrictionsS: NewRestrictionsService(tt.bk),
-		Clock:         clock,
+		TrustS:         caService,
+		TrustInternalS: caService,
+		PresenceS:      presenceService,
+		ProvisioningS:  NewProvisioningService(tt.bk),
+		WebS:           NewTestIdentityService(tt.bk),
+		Access:         NewAccessService(tt.bk),
+		EventsS:        eventsService,
+		ChangesC:       make(chan interface{}),
+		ConfigS:        configService,
+		LocalConfigS:   configService,
+		RestrictionsS:  NewRestrictionsService(tt.bk),
+		Clock:          clock,
 	}
 
 	return &tt
