+++ lib/services/local/trust.go
@@ -20,13 +20,16 @@ package local
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"log/slog"
 	"slices"
+	"sort"
 	"strings"
 
 	"github.com/gravitational/trace"
 
+	"github.com/gravitational/teleport/api/defaults"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/backend"
 	"github.com/gravitational/teleport/lib/services"
@@ -420,6 +423,419 @@ func (s *CA) UpdateUserCARoleMap(ctx context.Context, name string, roleMap types
 		return trace.Wrap(err)
 	}
 	return nil
+}
+
+// UpsertTrustedCluster creates or updates a TrustedCluster in the backend.
+func (s *CA) UpsertTrustedCluster(ctx context.Context, trustedCluster types.TrustedCluster) (types.TrustedCluster, error) {
+	if err := services.ValidateTrustedCluster(trustedCluster); err != nil {
+		return nil, trace.Wrap(err)
+	}
+	rev := trustedCluster.GetRevision()
+	value, err := services.MarshalTrustedCluster(trustedCluster)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	_, err = s.Put(ctx, backend.Item{
+		Key:      backend.Key(trustedClustersPrefix, trustedCluster.GetName()),
+		Value:    value,
+		Expires:  trustedCluster.Expiry(),
+		Revision: rev,
+	})
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	return trustedCluster, nil
+}
+
+// GetTrustedCluster returns a single TrustedCluster by name.
+func (s *CA) GetTrustedCluster(ctx context.Context, name string) (types.TrustedCluster, error) {
+	if name == "" {
+		return nil, trace.BadParameter("missing trusted cluster name")
+	}
+	item, err := s.Get(ctx, backend.Key(trustedClustersPrefix, name))
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	return services.UnmarshalTrustedCluster(item.Value, services.WithExpires(item.Expires), services.WithRevision(item.Revision))
+}
+
+// GetTrustedClusters returns all TrustedClusters in the backend.
+func (s *CA) GetTrustedClusters(ctx context.Context) ([]types.TrustedCluster, error) {
+	startKey := backend.ExactKey(trustedClustersPrefix)
+	result, err := s.GetRange(ctx, startKey, backend.RangeEnd(startKey), backend.NoLimit)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	out := make([]types.TrustedCluster, len(result.Items))
+	for i, item := range result.Items {
+		tc, err := services.UnmarshalTrustedCluster(item.Value,
+			services.WithExpires(item.Expires), services.WithRevision(item.Revision))
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+		out[i] = tc
+	}
+
+	sort.Sort(types.SortedTrustedCluster(out))
+	return out, nil
+}
+
+// DeleteTrustedCluster removes a TrustedCluster from the backend by name.
+func (s *CA) DeleteTrustedCluster(ctx context.Context, name string) error {
+	if name == "" {
+		return trace.BadParameter("missing trusted cluster name")
+	}
+	err := s.Delete(ctx, backend.Key(trustedClustersPrefix, name))
+	if err != nil {
+		if trace.IsNotFound(err) {
+			return trace.NotFound("trusted cluster %q is not found", name)
+		}
+	}
+	return trace.Wrap(err)
+}
+
+// UpsertTunnelConnection updates or creates tunnel connection
+func (s *CA) UpsertTunnelConnection(conn types.TunnelConnection) error {
+	if err := services.CheckAndSetDefaults(conn); err != nil {
+		return trace.Wrap(err)
+	}
+
+	rev := conn.GetRevision()
+	value, err := services.MarshalTunnelConnection(conn)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	_, err = s.Put(context.TODO(), backend.Item{
+		Key:      backend.Key(tunnelConnectionsPrefix, conn.GetClusterName(), conn.GetName()),
+		Value:    value,
+		Expires:  conn.Expiry(),
+		Revision: rev,
+	})
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	return nil
+}
+
+// GetTunnelConnection returns connection by cluster name and connection name
+func (s *CA) GetTunnelConnection(clusterName, connectionName string, opts ...services.MarshalOption) (types.TunnelConnection, error) {
+	item, err := s.Get(context.TODO(), backend.Key(tunnelConnectionsPrefix, clusterName, connectionName))
+	if err != nil {
+		if trace.IsNotFound(err) {
+			return nil, trace.NotFound("trusted cluster connection %q is not found", connectionName)
+		}
+		return nil, trace.Wrap(err)
+	}
+	conn, err := services.UnmarshalTunnelConnection(item.Value,
+		services.AddOptions(opts, services.WithExpires(item.Expires), services.WithRevision(item.Revision))...)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	return conn, nil
+}
+
+// GetTunnelConnections returns connections for a trusted cluster
+func (s *CA) GetTunnelConnections(clusterName string, opts ...services.MarshalOption) ([]types.TunnelConnection, error) {
+	if clusterName == "" {
+		return nil, trace.BadParameter("missing cluster name")
+	}
+	startKey := backend.ExactKey(tunnelConnectionsPrefix, clusterName)
+	result, err := s.GetRange(context.TODO(), startKey, backend.RangeEnd(startKey), backend.NoLimit)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	conns := make([]types.TunnelConnection, len(result.Items))
+	for i, item := range result.Items {
+		conn, err := services.UnmarshalTunnelConnection(item.Value,
+			services.AddOptions(opts, services.WithExpires(item.Expires), services.WithRevision(item.Revision))...)
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+		conns[i] = conn
+	}
+
+	return conns, nil
+}
+
+// GetAllTunnelConnections returns all tunnel connections
+func (s *CA) GetAllTunnelConnections(opts ...services.MarshalOption) ([]types.TunnelConnection, error) {
+	startKey := backend.ExactKey(tunnelConnectionsPrefix)
+	result, err := s.GetRange(context.TODO(), startKey, backend.RangeEnd(startKey), backend.NoLimit)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	conns := make([]types.TunnelConnection, len(result.Items))
+	for i, item := range result.Items {
+		conn, err := services.UnmarshalTunnelConnection(item.Value,
+			services.AddOptions(opts,
+				services.WithExpires(item.Expires),
+				services.WithRevision(item.Revision))...)
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+		conns[i] = conn
+	}
+
+	return conns, nil
+}
+
+// DeleteTunnelConnection deletes tunnel connection by name
+func (s *CA) DeleteTunnelConnection(clusterName, connectionName string) error {
+	if clusterName == "" {
+		return trace.BadParameter("missing cluster name")
+	}
+	if connectionName == "" {
+		return trace.BadParameter("missing connection name")
+	}
+	return s.Delete(context.TODO(), backend.Key(tunnelConnectionsPrefix, clusterName, connectionName))
+}
+
+// DeleteTunnelConnections deletes all tunnel connections for cluster
+func (s *CA) DeleteTunnelConnections(clusterName string) error {
+	if clusterName == "" {
+		return trace.BadParameter("missing cluster name")
+	}
+	startKey := backend.ExactKey(tunnelConnectionsPrefix, clusterName)
+	err := s.DeleteRange(context.TODO(), startKey, backend.RangeEnd(startKey))
+	return trace.Wrap(err)
+}
+
+// DeleteAllTunnelConnections deletes all tunnel connections
+func (s *CA) DeleteAllTunnelConnections() error {
+	startKey := backend.ExactKey(tunnelConnectionsPrefix)
+	err := s.DeleteRange(context.TODO(), startKey, backend.RangeEnd(startKey))
+	return trace.Wrap(err)
+}
+
+// CreateRemoteCluster creates remote cluster
+func (s *CA) CreateRemoteCluster(
+	ctx context.Context, rc types.RemoteCluster,
+) (types.RemoteCluster, error) {
+	value, err := json.Marshal(rc)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	item := backend.Item{
+		Key:     backend.Key(remoteClustersPrefix, rc.GetName()),
+		Value:   value,
+		Expires: rc.Expiry(),
+	}
+	lease, err := s.Create(ctx, item)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	rc.SetRevision(lease.Revision)
+	return rc, nil
+}
+
+// UpdateRemoteCluster updates selected remote cluster fields: expiry and labels
+// other changed fields will be ignored by the method
+func (s *CA) UpdateRemoteCluster(ctx context.Context, rc types.RemoteCluster) (types.RemoteCluster, error) {
+	if err := services.CheckAndSetDefaults(rc); err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	// Small retry loop to catch cases where there's a concurrent update which
+	// could cause conditional update to fail. This is needed because of the
+	// unusual way updates are handled in this method meaning that the revision
+	// in the provided remote cluster is not used. We should eventually make a
+	// breaking change to this behavior.
+	const iterationLimit = 3
+	for i := 0; i < iterationLimit; i++ {
+		existing, err := s.GetRemoteCluster(ctx, rc.GetName())
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+		existing.SetExpiry(rc.Expiry())
+		existing.SetLastHeartbeat(rc.GetLastHeartbeat())
+		existing.SetConnectionStatus(rc.GetConnectionStatus())
+		existing.SetMetadata(rc.GetMetadata())
+
+		updateValue, err := services.MarshalRemoteCluster(existing)
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+
+		lease, err := s.ConditionalUpdate(ctx, backend.Item{
+			Key:      backend.Key(remoteClustersPrefix, existing.GetName()),
+			Value:    updateValue,
+			Expires:  existing.Expiry(),
+			Revision: existing.GetRevision(),
+		})
+		if err != nil {
+			if trace.IsCompareFailed(err) {
+				// Retry!
+				continue
+			}
+			return nil, trace.Wrap(err)
+		}
+		existing.SetRevision(lease.Revision)
+		return existing, nil
+	}
+	return nil, trace.CompareFailed("failed to update remote cluster within %v iterations", iterationLimit)
+}
+
+// PatchRemoteCluster fetches a remote cluster and then calls updateFn
+// to apply any changes, before persisting the updated remote cluster.
+func (s *CA) PatchRemoteCluster(
+	ctx context.Context,
+	name string,
+	updateFn func(types.RemoteCluster) (types.RemoteCluster, error),
+) (types.RemoteCluster, error) {
+	// Retry to update the remote cluster in case of a conflict.
+	const iterationLimit = 3
+	for i := 0; i < 3; i++ {
+		existing, err := s.GetRemoteCluster(ctx, name)
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+
+		updated, err := updateFn(existing.Clone())
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+
+		switch {
+		case updated.GetName() != name:
+			return nil, trace.BadParameter("metadata.name: cannot be patched")
+		case updated.GetRevision() != existing.GetRevision():
+			// We don't allow revision to be specified when performing a patch.
+			// This is because it creates some complex semantics. Instead they
+			// should use the Update method if they wish to specify the
+			// revision.
+			return nil, trace.BadParameter("metadata.revision: cannot be patched")
+		}
+
+		updatedValue, err := services.MarshalRemoteCluster(updated)
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+
+		lease, err := s.ConditionalUpdate(ctx, backend.Item{
+			Key:      backend.Key(remoteClustersPrefix, name),
+			Value:    updatedValue,
+			Expires:  updated.Expiry(),
+			Revision: updated.GetRevision(),
+		})
+		if err != nil {
+			if trace.IsCompareFailed(err) {
+				// Retry!
+				continue
+			}
+			return nil, trace.Wrap(err)
+		}
+		updated.SetRevision(lease.Revision)
+		return updated, nil
+	}
+	return nil, trace.CompareFailed("failed to update remote cluster within %v iterations", iterationLimit)
+}
+
+// GetRemoteClusters returns a list of remote clusters
+// Prefer ListRemoteClusters. This will eventually be deprecated.
+// TODO(noah): REMOVE IN 17.0.0 - replace calls with ListRemoteClusters
+func (s *CA) GetRemoteClusters(ctx context.Context) ([]types.RemoteCluster, error) {
+	startKey := backend.ExactKey(remoteClustersPrefix)
+	result, err := s.GetRange(ctx, startKey, backend.RangeEnd(startKey), backend.NoLimit)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	clusters := make([]types.RemoteCluster, len(result.Items))
+	for i, item := range result.Items {
+		cluster, err := services.UnmarshalRemoteCluster(item.Value,
+			services.WithExpires(item.Expires),
+			services.WithRevision(item.Revision),
+		)
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+		clusters[i] = cluster
+	}
+	return clusters, nil
+}
+
+// ListRemoteClusters returns a page of remote clusters
+func (s *CA) ListRemoteClusters(
+	ctx context.Context, pageSize int, pageToken string,
+) ([]types.RemoteCluster, string, error) {
+	rangeStart := backend.Key(remoteClustersPrefix, pageToken)
+	rangeEnd := backend.RangeEnd(backend.ExactKey(remoteClustersPrefix))
+
+	// Adjust page size, so it can't be too large.
+	if pageSize <= 0 || pageSize > defaults.DefaultChunkSize {
+		pageSize = defaults.DefaultChunkSize
+	}
+
+	limit := pageSize + 1
+
+	result, err := s.GetRange(ctx, rangeStart, rangeEnd, limit)
+	if err != nil {
+		return nil, "", trace.Wrap(err)
+	}
+
+	clusters := make([]types.RemoteCluster, 0, len(result.Items))
+	for _, item := range result.Items {
+		cluster, err := services.UnmarshalRemoteCluster(item.Value,
+			services.WithExpires(item.Expires),
+			services.WithRevision(item.Revision),
+		)
+		if err != nil {
+			slog.WarnContext(ctx, "Skipping item during ListRemoteClusters because conversion from backend item failed", "key", item.Key, "error", err)
+			continue
+		}
+		clusters = append(clusters, cluster)
+	}
+
+	next := ""
+	if len(clusters) > pageSize {
+		next = backend.GetPaginationKey(clusters[pageSize])
+		clear(clusters[pageSize:])
+		// Truncate the last item that was used to determine next row existence.
+		clusters = clusters[:pageSize]
+	}
+	return clusters, next, nil
+}
+
+// GetRemoteCluster returns a remote cluster by name
+func (s *CA) GetRemoteCluster(
+	ctx context.Context, clusterName string,
+) (types.RemoteCluster, error) {
+	if clusterName == "" {
+		return nil, trace.BadParameter("missing parameter cluster name")
+	}
+	item, err := s.Get(ctx, backend.Key(remoteClustersPrefix, clusterName))
+	if err != nil {
+		if trace.IsNotFound(err) {
+			return nil, trace.NotFound("remote cluster %q is not found", clusterName)
+		}
+		return nil, trace.Wrap(err)
+	}
+	rc, err := services.UnmarshalRemoteCluster(item.Value,
+		services.WithExpires(item.Expires),
+		services.WithRevision(item.Revision),
+	)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	return rc, nil
+}
+
+// DeleteRemoteCluster deletes remote cluster by name
+func (s *CA) DeleteRemoteCluster(
+	ctx context.Context, clusterName string,
+) error {
+	if clusterName == "" {
+		return trace.BadParameter("missing parameter cluster name")
+	}
+	return s.Delete(ctx, backend.Key(remoteClustersPrefix, clusterName))
+}
+
+// DeleteAllRemoteClusters deletes all remote clusters
+func (s *CA) DeleteAllRemoteClusters(ctx context.Context) error {
+	startKey := backend.ExactKey(remoteClustersPrefix)
+	err := s.DeleteRange(ctx, startKey, backend.RangeEnd(startKey))
+	return trace.Wrap(err)
 }
 
 // catToItem builds a backend.Item corresponding to the supplied CA.
